+++ app/src/composables/unsaved-changes/index.ts
@@ -1,0 +1,4 @@
+import unsavedChanges from './unsaved-changes';
+
+export { unsavedChanges };
+export default unsavedChanges;
